requirements.txt
fastapi
uvicorn[standard]
python-dotenv
aiohttp
PyMuPDF
python-docx
google-generativeai
pydantic
